feat: Add function to format dates to quarter periods.
This commit adds a new JavaScript function, formatDateToQuarter, to the project.
The function takes a fiscal date as input and formats it into a quarter string.
It is designed to be used with the map function to format an array of fiscal dates
into an array of corresponding quarters.

This functionality enables easier organization and analysis of financial data,
as it provides a standardized representation of fiscal dates in quarters.