feat(component) Sidebar component
The Sidebar component enables users to navigate effortlessly between pages within the application. Additionally, the component includes a "simplifiedSidebar" boolean that can be passed in as a prop, enhancing the appearance of the sidebar to occupy less width. This feature enhances the responsiveness of our pages while maintaining the sidebar's presence.